feat: 🎸 improve CombinedLinkActions styles
Makes actions bold and removes underline text decoration.
